Refactor authority / interaction flow
Previously, the code was quite difficult to understand as two distinct
paths (/licence/council && /licence/council/interaction) were rendered
by one action and shared the same view. It was quite difficult to work
out what belonged to which path. Now we have a disctint controller
action for the interactions flow with it's own view.

We also now use a base_authority view partial to keep the code DRY.